refactor: Extract and cleanup eth transact function
